+++ src/pokemon_summary_screen.c
@@ -103,7 +103,7 @@ void HandleDuplicateNames_SummaryScreen()
     u8 firstType = gBaseStats[species].type1;
     u8 ateType, terrainType;
 
-    for(u8 i = 0; i < ARRAY_COUNT(sDuplicateAbilities); i++)
+    for(u16 i = 0; i < ARRAY_COUNT(sDuplicateAbilities); i++)
     {
         if(ability == sDuplicateAbilities[i].currAbility && species == sDuplicateAbilities[i].species)
         {
